Remove compat example
Because we are `yarn link`ing, this example is not actually testing
what we cared about (which was the current CLI could work with 0.10.0
versions of Pulumi). We could get this test working again, but there
are many versions of nodes it would not support, so lets just delete
it. The coverage we got isn't really interesting anymore.